requirements.txt
# Samsung PRISM - Harmful Content Detection Pipeline
# Requirements for Google Colab (GPU-enabled)

# Core ML Frameworks
ultralytics>=8.0.200          # YOLOv8 for object detection
torch>=2.0.0                   # PyTorch deep learning
torchvision>=0.15.0           # Vision utilities
transformers>=4.35.0          # DistilBERT for NLP

# OCR
easyocr>=1.7.1                # Text extraction from images

# Image/Video Processing
opencv-python>=4.8.0          # Image processing
Pillow>=10.0.0                # Image handling
ffmpeg-python>=0.2.0          # Video frame extraction

# Data Handling
numpy>=1.24.0                 # Numerical operations
pandas>=2.0.0                 # Data manipulation
pyyaml>=6.0                   # Config file parsing

# Visualization
matplotlib>=3.7.0             # Plotting
seaborn>=0.12.0               # Statistical visualization

# Dataset Management
roboflow>=1.1.0               # Dataset augmentation (optional)
albumentations>=1.3.0         # Image augmentation

# Utilities
tqdm>=4.65.0                  # Progress bars
scikit-learn>=1.3.0           # Metrics and evaluation
